move: 로그인, 회원가입 스타일컴포넌트 이동
