+++ src/components/UI/SignInput.jsx
@@ -2,7 +2,7 @@ import { useEffect, useRef } from "react";
 import styled from "styled-components";
 
 const SignInput = ({ inputs, setInputs, label, name, type, placeholder, terms, regex, firstFocus }) => {
-  // input들을 제어하는 함수
+  // input을 제어하는 함수
   const inputOnChange = (event) => {
     const { name, value } = event.target;
     setInputs((prev) => {
