+++ src/pages/public/SignIn.jsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
-import SignInput from "../../components/SignInputs";
+import SignInput from "../../components/UI/SignInput";
 import { useNavigate } from "react-router-dom";
 import useSignInHandler from "../../hooks/useSignInHandler";
-import { NavGuide, SignBtn, SignFrom, SignNav } from "../../styles/SignStyles";
+import { NavGuide, SignBtn, SignFrom, SignNav } from "../../components/UI/SignStyles";
 import styled from "styled-components";
 import { EMAIL_INPUT_REGEX, EMAIL_REGEX, PASSWORD_INPUT_REGEX, PASSWORD_REGEX } from "../../constant/regularExpression";
 import { Toaster } from "sonner";
